Splits TUI to own file.
