+++ src/beepy_web_radio/main.py
@@ -1,21 +1,5 @@
-from textual.app import App, ComposeResult
-from textual.binding import Binding
-from textual.widgets import Footer, Label
-
 from .cli import main as cli_main
-
-
-class BeepyWebRadioApp(App):
-    BINDINGS = [
-        Binding("q", "quit", "Quit", show=True),
-    ]
-
-    def compose(self) -> ComposeResult:
-        yield Label("Beepy Web Radio")
-        yield Footer()
-
-    def on_mount(self) -> None:
-        pass
+from .tui import BeepyWebRadioApp
 
 
 def main() -> None:
